git: Fix bug with locks in generate_tags()
Because the line that actually generates tags is run as a background
process, we used to go right ahead and proceed to the next line (which
removes the lock file) while the tags were still being generated. Now,
we run the entirety of `generate_tags()` in a background process so the
tag generation line blocks and we don't remove the lock file until after
it completes.