nvim: Don't open horizontal split terminals across entire bottom
Now they'll open below the current split but with the same width as the
current split. Most of the time when I'm opening a horizontal terminal
it's just to do something quick that's tangential to what I'm doing in
my existing, vertical split terminal so I want it grouped with that one
and just under it. So it makes sense to ditch `:botright` here but keep
it for the vertical terminal keymaps.