fish: Add vendor completions to `fish_complete_path`
The old method of sourcing completions in config.fish meant that all the
completions would be loaded any time a fish process was launched. This
is inefficient because I might not end up using those completions. It's
better to defer loading the completion until I actually use one of those
commands. This is (I believe) what fish does normally for the
completions that it ships with. The `fish_complete_path` variable can be
used to add directories that fish will search for completions in. So by
adding the vendor completions directory to `fish_complete_path`, I can
achieve the same lazy loading behavior for vendor completions too. \o/