vim: Fix bug where random keys are entered when launching neovim
Something changed when I upgraded to neovim v0.11 and now launching a
new neovim process always inserts some characters into the buffer. It
appears that some keys are being automatically triggered. There are some
relevant open issues on the neovim issue tracker. The only solution I
can find that works is to remove `nottimeout` and replace it with
`ttimeoutlen=0`. In the absence of further clarity, this'll have to
suffice.