nvim: Auto-map <C-h> to neoterm's "run last command in terminal"
Before I had to do this manually in each neovim process which was a
pain. Now I don't have to!

`VimEnter` is a bit of a hack but the `Tmap` command isn't available at
the time that the init file is being evaluated so we need to wait a bit.